Remove suffix for extractor-v2
