+++ packages/core-generators/src/renderers/typescript/extractor/render-ts-import-providers.ts
@@ -1,4 +1,4 @@
-import type { TemplateExtractorTemplateEntry } from '@baseplate-dev/sync/extractor-v2';
+import type { TemplateExtractorTemplateEntry } from '@baseplate-dev/sync';
 
 import { quot } from '@baseplate-dev/utils';
 import { camelCase } from 'change-case';
