+++ packages/sync/package.json
@@ -20,13 +20,9 @@
       "types": "./dist/index.d.ts",
       "import": "./dist/index.js"
     },
-    "./extractor-v2": {
-      "types": "./dist/templates/extractor-v2/index.d.ts",
-      "import": "./dist/templates/extractor-v2/index.js"
-    },
-    "./extractor-v2/test-utils": {
-      "types": "./dist/templates/extractor-v2/test-utils/index.d.ts",
-      "import": "./dist/templates/extractor-v2/test-utils/index.js"
+    "./extractor-test-utils": {
+      "types": "./dist/templates/extractor/test-utils/index.d.ts",
+      "import": "./dist/templates/extractor/test-utils/index.js"
     }
   },
   "files": [
