+++ packages/sync/src/templates/extractor/configs/extractor-config.schema.ts
