+++ packages/sync/src/templates/extractor/configs/index.ts
