+++ packages/sync/src/templates/extractor/configs/providers-config.schema.ts
