+++ packages/sync/src/templates/extractor/configs/template-extractor-config-lookup.unit.test.ts
