+++ packages/sync/src/templates/extractor/configs/try-create-extractor-json.ts
