+++ packages/sync/src/templates/extractor/constants/directories.ts
