+++ packages/sync/src/templates/extractor/index.ts
