+++ packages/sync/src/templates/extractor/run-template-file-extractors.ts
