+++ packages/sync/src/templates/extractor/runner/initialize-template-extractor-plugins.ts
