+++ packages/sync/src/templates/extractor/runner/template-extractor-file-container.ts
