+++ packages/sync/src/templates/extractor/runner/template-extractor-plugin.ts
