+++ packages/sync/src/templates/extractor/runner/template-file-extractor.ts
