+++ packages/sync/src/templates/extractor/test-utils/file-system-utils.ts
