+++ packages/sync/src/templates/extractor/utils/cleanup-unused-template-files.ts
