+++ packages/sync/src/templates/extractor/utils/cleanup-unused-template-files.unit.test.ts
