+++ packages/sync/src/templates/extractor/utils/index.ts
