+++ packages/sync/src/templates/extractor/utils/write-extractor-template-jsons.ts
