+++ packages/sync/src/templates/index.ts
@@ -1,4 +1,5 @@
 export * from './constants.js';
+export * from './extractor/index.js';
 export * from './metadata/index.js';
 export type * from './types.js';
 export * from './utils/index.js';
